\name{relex-package}
\alias{relex-package}
\alias{relex-package}
\title{Linear models for proteomics data.}
\description{Linear models for proteomics data.
This package provides a set of methods for import/export, normalisation and
statistical analysis of MS/MS proteomics data that has been processed using
RelEx.  I intend to support more types of proteomics data, so if you have
any other type of data that you'd like incorporated, please let the package
maintainer know.}
\details{Starting with relex data that you have generated, go to \code{\link{import.relex.experiment}},
MORE INFO TO FOLLOW

\tabular{ll}{
Package: \tab relex\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-07-16\cr
License: \tab GPL2\cr
LazyLoad: \tab yes\cr
}}
\alias{relex-package}
\alias{relex}
\docType{package}
\author{Mark Cowley

Maintainer: Mark Cowley <m.cowley@garvan.org.au>}
\seealso{\code{\link{import.relex.experiment}}, \code{\link{normalize.relex.experiment}}, \code{\link[limma]{limma}}}
\references{L. Ting, M.J. Cowley, S.L. Hoon, M. Guilhaus, M.J. Raftery & R.
Cavicchioli. 2009. Normalization and Statistical Analysis of Quantitative
Proteomics Data Generated by Metabolic Labeling. Mol Cell Proteomics 8
(Oct. 2009), 2227-2242.}
\keyword{package}
\examples{# much needed examples go here.}

